Fixed Wing's mess teehee

+++ src/main/java/JBKRMobile/Buy.java
@@ -1,7 +1,7 @@
 package JBKRMobile;
 
 public class Buy extends Transaction {
-    public Buy(String date, String ticker, long quantity, double price) {
+    public Buy(String date, String ticker, int quantity, double price) {
         super(date, ticker, quantity, price);
     }
